Make client Swarm-specific
