+++ package.json
@@ -19,7 +19,7 @@
     "test:ci": "jest --ci --config jest.config.ci.js --runInBand",
     "test": "yarn test:flow && yarn test:project",
     "test:all": "yarn lint && yarn test",
-    "build": "lerna run build && cd packages/browser && yarn build:dist",
+    "build": "lerna run build && cd packages/swarm-browser && yarn build:dist",
     "start": "yarn test && yarn build"
   },
   "devDependencies": {
