--- packages/browser/README.md
@@ -1,17 +1,0 @@
-# @erebos/browser
-
-Erebos library for the browser - [MainframeHQ/erebos](https://github.com/MainframeHQ/erebos).
-
-## Installation
-
-```sh
-yarn add @erebos/browser
-```
-
-## API
-
-See the [library documentation](../../docs/README.md).
-
-## License
-
-MIT
