+++ packages/cli/package.json
@@ -28,7 +28,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.0.0",
-    "@erebos/node": "^0.3.0",
+    "@erebos/swarm-node": "^0.3.0",
     "@oclif/command": "^1.5.0",
     "@oclif/config": "^1.7.3",
     "@oclif/plugin-help": "^2.1.0",
