--- packages/erebos/README.md
@@ -1,19 +1,0 @@
-# erebos
-
-Universal Erebos library - [MainframeHQ/erebos](https://github.com/MainframeHQ/erebos).
-
-Uses [@erebos/browser](../browser) or [@erebos/node](../node) depending on the environment. You should use one of these packages directly if you do not need to support both environment.
-
-## Installation
-
-```sh
-yarn add erebos
-```
-
-## API
-
-See the [library documentation](../../docs/README.md).
-
-## License
-
-MIT
