--- packages/node/README.md
@@ -1,17 +1,0 @@
-# @erebos/node
-
-Erebos library for the node - [MainframeHQ/erebos](https://github.com/MainframeHQ/erebos).
-
-## Installation
-
-```sh
-yarn add @erebos/node
-```
-
-## API
-
-See the [library documentation](../../docs/README.md).
-
-## License
-
-MIT
