+++ packages/swarm-browser/.babelrc
