+++ packages/swarm-browser/README.md
@@ -1,0 +1,17 @@
+# @erebos/swarm-browser
+
+Erebos Swarm library for the browser - [MainframeHQ/erebos](https://github.com/MainframeHQ/erebos).
+
+## Installation
+
+```sh
+yarn add @erebos/swarm-browser
+```
+
+## API
+
+See the [library documentation](../../docs/README.md).
+
+## License
+
+MIT
