+++ packages/swarm-browser/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@erebos/browser",
+  "name": "@erebos/swarm-browser",
   "version": "0.3.0",
-  "description": "Ethereum and Swarm client for browser",
+  "description": "Swarm client for browser",
   "repository": "git@github.com:MainframeHQ/erebos.git",
   "main": "cjs/index.js",
   "module": "esm/index.js",
@@ -28,13 +28,10 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0",
     "@erebos/api-bzz-browser": "^0.3.0",
-    "@erebos/api-eth": "^0.3.0",
-    "@erebos/api-net": "^0.3.0",
     "@erebos/api-pss": "^0.3.0",
-    "@erebos/api-shh": "^0.3.0",
-    "@erebos/api-web3": "^0.3.0",
     "@erebos/client-base": "^0.3.0",
     "@mainframe/rpc-browser": "^0.2.1",
+    "@mainframe/rpc-stream": "^0.2.4",
     "@mainframe/utils-hex": "^0.1.1",
     "rxjs": "^6.3.0"
   },
