+++ packages/swarm-node/README.md
@@ -1,0 +1,17 @@
+# @erebos/swarm-node
+
+Erebos Swarm library for node - [MainframeHQ/erebos](https://github.com/MainframeHQ/erebos).
+
+## Installation
+
+```sh
+yarn add @erebos/swarm-node
+```
+
+## API
+
+See the [library documentation](../../docs/README.md).
+
+## License
+
+MIT
