+++ packages/swarm-node/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@erebos/node",
+  "name": "@erebos/swarm-node",
   "version": "0.3.0",
-  "description": "Ethereum and Swarm client for node",
+  "description": "Swarm client for node",
   "repository": "git@github.com:MainframeHQ/erebos.git",
   "main": "lib/index.js",
   "author": "Mainframe",
@@ -19,11 +19,7 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0",
     "@erebos/api-bzz-node": "^0.3.0",
-    "@erebos/api-eth": "^0.3.0",
-    "@erebos/api-net": "^0.3.0",
     "@erebos/api-pss": "^0.3.0",
-    "@erebos/api-shh": "^0.3.0",
-    "@erebos/api-web3": "^0.3.0",
     "@erebos/client-base": "^0.3.0",
     "@mainframe/rpc-node": "^0.2.1",
     "@mainframe/rpc-stream": "^0.2.4",
