+++ packages/swarm/README.md
@@ -1,0 +1,19 @@
+# @erebos/swarm
+
+Universal Erebos Swarm library - [MainframeHQ/erebos](https://github.com/MainframeHQ/erebos).
+
+Uses [@erebos/swarm-browser](../swarm-browser) or [@erebos/swarm-node](../swarm-node) depending on the environment. You should use one of these packages directly if you do not need to support both environment.
+
+## Installation
+
+```sh
+yarn add @erebos/swarm
+```
+
+## API
+
+See the [library documentation](../../docs/README.md).
+
+## License
+
+MIT
